[2/2] Split off Download- from RepositoryController
Move split-off class to its own module
