+++ test/test_doctests.py
@@ -41,9 +41,9 @@ import azul.plugins.metadata.hca.service.contributor_matrices
 import azul.plugins.repository.canned
 import azul.plugins.repository.tdr_hca
 import azul.service.app_controller
+import azul.service.download_controller
 import azul.service.drs_controller
 import azul.service.manifest_service
-import azul.service.repository_controller
 import azul.strings
 import azul.terra
 import azul.terraform
@@ -99,9 +99,9 @@ def load_tests(_loader,
         azul.plugins.repository.tdr_hca,
         azul.plugins.metadata.hca.indexer.transform,
         azul.service.app_controller,
+        azul.service.download_controller,
         azul.service.drs_controller,
         azul.service.manifest_service,
-        azul.service.repository_controller,
         azul.strings,
         azul.terra,
         azul.terraform,
